import-to-bq/requirements.txt
google-cloud-bigquery==3.3.2
google-cloud-storage==2.5.0
functions-framework==3.2.0
pandas==1.4.4